signout api

+++ admin-front-app/src/actions/auth.action.js
@@ -60,9 +60,18 @@ export const isUSerLoggedIn = () => {
 
 export const signout = () => {
     return async dispatch => {
-        localStorage.clear();
-        dispatch({
-            type: authConstants.LOGOUT_REQUEST
-        });
+
+        dispatch({ type: authConstants.LOGOUT_REQUEST })
+        const res = await axiosInstance.post('/admin/signout')
+        if (res.status === 200) {
+            localStorage.clear();
+            dispatch({ type: authConstants.LOGOUT_SUCCESS });
+        } else {
+            dispatch({
+                type: authConstants.LOGOUT_FAILURE,
+                payload: { erorr: res.data.error }
+            });
+        }
+
     }
 }
